Alignment tweeks
Center the title, lead <p> and button on the blog page.
Add padding to the title, lead <p> and button on the blog page.
Change the blog post's title font-size on the blog page.
Add padding to the button on the post page.
Shift the navbar links to the right.